+++ src/game/scenes/screen/interface/player-hud/skills/index.tsx
@@ -1,28 +1,73 @@
-import { useGame } from 'phaser-react-ui';
-import React, { useEffect, useState } from 'react';
+import { useGame, useMobilePlatform } from 'phaser-react-ui';
+import React, {
+  useEffect, useMemo, useRef, useState,
+} from 'react';
 
+import { PLAYER_SKILLS } from '~const/world/entities/player';
 import { Button } from '~scene/system/interface/button';
 import { Hint } from '~scene/system/interface/hint';
 import { IGame } from '~type/game';
 import { TutorialStep } from '~type/tutorial';
+import { PlayerSkill } from '~type/world/entities/player';
 
-import { UpgradesList } from './list';
-import { Wrapper } from './styles';
+import { UpgradesListItem } from './item';
+import { Wrapper, Container } from './styles';
 
 export const Skills: React.FC = () => {
   const game = useGame<IGame>();
 
+  const isMobile = useMobilePlatform();
+
+  const upgradeTypes = useMemo(
+    () => Object.keys(PLAYER_SKILLS) as PlayerSkill[],
+    [],
+  );
+
   const [isOpened, setOpened] = useState(false);
   const [hint, setHint] = useState(false);
 
-  const onClickButton = (event: React.MouseEvent<HTMLDivElement>) => {
+  const refContainer = useRef<HTMLDivElement>(null);
+
+  const onClickButton = () => {
     setOpened(!isOpened);
-    event.stopPropagation();
   };
 
   const onClose = () => {
     setOpened(false);
   };
+
+  const onClickOutside = (event: MouseEvent | TouchEvent) => {
+    if (!refContainer.current) {
+      return;
+    }
+
+    const isInside = event.composedPath().includes(refContainer.current);
+
+    if (!isInside) {
+      onClose();
+    }
+  };
+
+  const onKeyPress = (event: KeyboardEvent) => {
+    if (event.key === 'Escape') {
+      onClose();
+      event.stopPropagation();
+    }
+  };
+
+  useEffect(() => {
+    if (!isOpened) {
+      return;
+    }
+
+    document.addEventListener(isMobile ? 'touchend' : 'click', onClickOutside);
+    document.addEventListener('keyup', onKeyPress);
+
+    return () => {
+      document.removeEventListener(isMobile ? 'touchend' : 'click', onClickOutside);
+      document.removeEventListener('keyup', onKeyPress);
+    };
+  }, [isOpened]);
 
   useEffect(
     () => game.tutorial.bind(TutorialStep.UPGRADE_SKILL, {
@@ -33,15 +78,22 @@ export const Skills: React.FC = () => {
   );
 
   return (
-    <Wrapper>
+    <Wrapper ref={refContainer}>
       <Button onClick={onClickButton} view={isOpened ? 'active' : undefined}>
         SKILLS
       </Button>
-      {isOpened && <UpgradesList onClose={onClose} />}
-      {hint && !isOpened && (
-        <Hint side="top" align="left">
-          Click to upgrade skills
-        </Hint>
+      {isOpened ? (
+        <Container>
+          {upgradeTypes.map((type) => (
+            <UpgradesListItem key={type} type={type} />
+          ))}
+        </Container>
+      ) : (
+        hint && (
+          <Hint side="top" align="left">
+            Click to upgrade skills
+          </Hint>
+        )
       )}
     </Wrapper>
   );
